Fix warnings in compile

+++ roboteam_ai/src/bt/tactics/HaltTactic.cpp
@@ -33,7 +33,7 @@ void HaltTactic::setName(std::string newName) {
 void HaltTactic::initialize() {
     std::vector<std::string> roleNames = {"1", "2", "3", "4",
                                           "5", "6", "7", "8"};
-    while (claimedRobots < roleNames.size()) {
+    while (claimedRobots < static_cast<int>(roleNames.size())) {
         robotIDs.insert(dealer::claimRobotForTactic(robotType::random, name, roleNames[claimedRobots]));
         if (robotIDs.find(-1) == robotIDs.end()) {
             claimedRobots++;
